Rename files to tconfig
